修复：消除完整构建中剩余的全部编译警告，保持 Qt5.14+/Qt6 兼容
QButtonGroup 废弃信号（buttonClicked(int)/buttonPressed(int) 自 5.15 废弃、Qt6 移除）：
- 版本分界从 6.0 改为 5.15：>=5.15 用 idClicked/idPressed（since 5.15），5.14 回退 QOverload 形式（5.14 下该重载未废弃，无警告）
- 涉及 DANodeItemSettingWidget、DAChartAxisSettingPanel、DAChartItemSettingPanel(2处)、DASettingPageGeneral
- 同步修正 DAGlobals.h 中 Qt5Qt6Compat_Connect_ButtonGroupClicked_int 兼容宏的分界

QComboBox::currentIndexChanged(const QString&) 废弃（since 5.15）：
- DADataManagerComboBox 改用 currentTextChanged（since Qt 5.0，原 Qt6 分支已在用），去除版本分支

全局 endl 操纵符废弃（QTextStreamFunctions::endl since 5.15）：
- DAPluginOption、DAPluginManager(3处) 统一改用 Qt::endl（since Qt 5.14，与项目最低版本一致），去除版本分支

C4267 size_t→int 收窄：
- DAChartAddtGridRasterDataWidget、DAStatsPlotCoordinator 的 reserve() 显式 static_cast<int>

验证：Release 与 Debug 双配置全量构建均零警告零错误

涉及文件：
- src/DAGlobals.h
- src/DAGui/DANodeItemSettingWidget.cpp
- src/DAGui/ChartSetting/DAChartAxisSettingPanel.cpp
- src/DAGui/ChartSetting/DAChartItemSettingPanel.cpp
- src/APP/SettingPages/DASettingPageGeneral.cpp
- src/DAGui/DADataManagerComboBox.cpp
- src/DAPluginSupport/DAPluginOption.cpp
- src/DAPluginSupport/DAPluginManager.cpp
- src/DAGui/ChartAddItem/DAChartAddtGridRasterDataWidget.cpp
- src/APP/DAStatsPlotCoordinator.cpp

+++ src/APP/DAStatsPlotCoordinator.cpp
@@ -703,7 +703,7 @@ void DAStatsPlotCoordinator::plotBarplot(const QJsonObject& params,
             auto groupVals = pybind11::cast<pybind11::list>(groupValsObj);
             DA::DABarChartData groupBarData;
             groupBarData.categories = categories;
-            groupBarData.values.reserve(groupVals.size());
+            groupBarData.values.reserve(static_cast< int >(groupVals.size()));
             for (auto val : groupVals) {
                 if (!val.is_none()) {
                     groupBarData.values.append(pybind11::cast<double>(val));
